+++ packages/miniapp-builder-shared/src/constants.js
@@ -3,4 +3,6 @@ module.exports = {
   WECHAT_MINIPROGRAM: 'wechat-miniprogram',
   BYTEDANCE_MICROAPP: 'bytedance-microapp',
   QUICKAPP: 'quickapp',
+  BAIDU_SMARTPROGRAM: 'baidu-smartprogram',
+  KUAISHOU_MINIPROGRAM: 'kuaishou-miniprogram'
 };