+++ packages/miniapp-render/src/node/node.js
@@ -1,3 +1,5 @@
+// eslint-disable-next-line import/no-extraneous-dependencies
+import { isBaiduSmartProgram } from 'universal-env';
 import EventTarget from '../event/event-target';
 import { getId } from '../utils/tool';
 import cache from '../utils/cache';
@@ -37,7 +39,8 @@ class Node extends EventTarget {
 
   get _path() {
     if (this.parentNode !== null) {
-      const index = '[' + this.parentNode.childNodes.indexOf(this) + ']';
+      const childIndex = this.parentNode.childNodes.indexOf(this);
+      const index = isBaiduSmartProgram ? childIndex : `[${childIndex}]`;
 
       return `${this.parentNode._path}.children.${index}`;
     }