+++ packages/miniapp-builder-shared/CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## [0.2.4] - 2021-05-19
+
+### Changed
+
+- Update kuaishou miniprogram wxs extension name
+
+## [0.2.4] - 2021-05-10
+
+### Added
+
+- Add baidu and kuaishou config
+
+## [0.2.3] - 2021-04-27
+
+### Changed
+
+- Update homepage url in `package.json`
+
 ## [0.2.2] - 2021-03-30
 
 ### Added